Move doc-strings.ts files into tests folders
