+++ packages/nimble-components/src/button/tests/button.stories.ts
@@ -9,7 +9,7 @@ import {
     contentHiddenDescription,
     endIconDescription,
     iconDescription
-} from '../../patterns/button/doc-strings';
+} from '../../patterns/button/tests/doc-strings';
 
 interface ButtonArgs {
     label: string;
